Namespace changes and file reorganization
Changed some namespaces to keep things uniform and moving some project
folders with the intention of making them common between the visual
studio extension and a command line project. Planning to add a command
line project that does C# to TypeScript generation to avoid the overhead
of having to launch the Visual Studio experimental instance to test code
changes.